RentalUI updated and done

+++ src/GUILayer/Rental/Item/DeleteUI.java
@@ -8,20 +8,24 @@ import javax.swing.JOptionPane;
 import javax.swing.JTextField;
 import javax.swing.JButton;
 
+import ControlLayer.LeaseCtrl;
 import GUILayer.GlobalUI;
 
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
+import java.awt.event.KeyAdapter;
+import java.awt.event.KeyEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 
-public class DeleteUI {
+public class DeleteUI 
+{
 
 	private static JFrame _frame;
 	private static DeleteUI _instance;
 	private JPanel contentPane;
 	private JTextField txtBarcode;
-
+	private LeaseCtrl _leaseCtrl;
 	public static JFrame createWindow()
 	{
 		if(_instance == null)
@@ -30,7 +34,9 @@ public class DeleteUI {
 		return _frame;
 	}
 	
-	private DeleteUI() {
+	private DeleteUI() 
+	{
+		_leaseCtrl = new LeaseCtrl();
 		_frame = new JFrame();
 		_frame.setResizable(false);
 		_frame.setTitle("Slet udlejningsprodukt");
@@ -42,19 +48,28 @@ public class DeleteUI {
 		_frame.setContentPane(contentPane);
 		contentPane.setLayout(null);
 		
-		JLabel lblBarCode = new JLabel("Stregkode");
-		lblBarCode.setBounds(10, 11, 116, 17);
+		JLabel lblBarCode = new JLabel("Produkt nummer");
+		lblBarCode.setBounds(10, 11, 131, 17);
 		contentPane.add(lblBarCode);
 		
 		txtBarcode = new JTextField();
-		txtBarcode.setBounds(120, 8, 314, 20);
+		txtBarcode.addKeyListener(new KeyAdapter() {
+			public void keyReleased(KeyEvent e) {
+				if(txtBarcode.getText().length() > 0)
+				{
+					GlobalUI.checkIfLong(txtBarcode);
+				}
+			}
+		});
+		txtBarcode.setBounds(147, 8, 287, 20);
 		contentPane.add(txtBarcode);
 		txtBarcode.setColumns(10);
 		
-		JButton btnCreate = new JButton("Udf\u00F8r");
+		JButton btnCreate = new JButton("Slet");
 		btnCreate.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				deleteItem();
+				if(txtBarcode.getText().length() > 0)
+					deleteItem();
 			}
 		});
 		btnCreate.setBounds(193, 66, 117, 23);
@@ -78,21 +93,27 @@ public class DeleteUI {
 		});
 	}
 	
-	private void deleteItem() {
+	private void deleteItem() 
+	{
 		boolean succeeded = false;
 		
-		try {	
-			if(succeeded) {
+		try 
+		{	long itemNumber = Long.parseLong(txtBarcode.getText());
+			succeeded = _leaseCtrl.deleteLeaseItem(itemNumber);
+			if(succeeded) 
+			{
 				JOptionPane.showMessageDialog(null, GlobalUI.messageHandling(15), "INFORMATION!", JOptionPane.INFORMATION_MESSAGE);
 				_instance = null;
 				_frame.dispose();
 			}
-			else {
+			else 
+			{
 				JOptionPane.showMessageDialog(null, GlobalUI.messageHandling(16), "FEJL!", JOptionPane.WARNING_MESSAGE);
 			}
 			
 		}
-		catch (Exception e) {
+		catch (Exception e) 
+		{
 			JOptionPane.showMessageDialog(null, GlobalUI.messageHandling(99), "FEJL!", JOptionPane.WARNING_MESSAGE);
 		}
 	}
